+++ includes/JsonStructure.hpp
