+++ lib/JsonItem.cpp
@@ -1,4 +1,4 @@
-#include "JsonItem.hpp"
+#include "../includes/JsonItem.hpp"
 
 namespace jsoncpp {
