+++ lib/JsonValue.cpp
@@ -1,4 +1,4 @@
-#include "JsonValue.hpp"
+#include "../includes/JsonValue.hpp"
 
 namespace jsoncpp {
